pyproject.toml
[project]
name = "uvm-burger-shop"
version = "0.1.0"
description = "UVM Burger Shop - Interactive Learning Platform"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "python-frontmatter>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

